Extract compile command from cli/index.ts
