+++ cli/commands/index.ts
@@ -1,3 +1,4 @@
+export { compileCommand } from "df/cli/commands/compile_command";
 export { formatCommand } from "df/cli/commands/format_command";
 export { helpCommand } from "df/cli/commands/help_command";
 export { initCommand } from "df/cli/commands/init_command";
