+++ cli/index.ts
@@ -1,15 +1,15 @@
-import * as chokidar from "chokidar";
 import * as fs from "fs";
 import parseDuration from "parse-duration";
 import * as path from "path";
 import yargs from "yargs";
 
-import { build, compile, credentials, prune, run, test } from "df/cli/api";
+import { build, compile, credentials, run, test } from "df/cli/api";
 import { BigQueryDbAdapter } from "df/cli/api/dbadapters/bigquery";
 import { LineageEmitter } from "df/cli/api/lineage/emitter";
 import { createLineageEmitter as createLineageEmitterFromFactory } from "df/cli/api/lineage/emitter_factory";
 import { prettyJsonStringify } from "df/cli/api/utils";
 import {
+  compileCommand,
   formatCommand,
   helpCommand,
   initCommand,
@@ -25,14 +25,13 @@ import {
   projectDirMustExistOption,
   projectDirOption,
   quietCompileOption,
+  requiresSelection,
   splitCommas,
   timeoutOption
 } from "df/cli/common_options";
 import {
-  compiledGraphOutputType,
   Logger,
   print,
-  printCompiledGraph,
   printCompiledGraphErrors,
   printError,
   printExecutedAction,
@@ -48,8 +47,6 @@ import {
 import { createYargsCli, INamedOption } from "df/cli/yargswrapper";
 import { targetAsReadableString } from "df/core/targets";
 import { dataform } from "df/protos/ts";
-
-const RECOMPILE_DELAY = 500;
 
 // Maximum time to wait for outstanding lineage emissions to complete before
 // `dataform run` returns. Lineage emission is fail-open — if we don't drain
@@ -70,19 +67,6 @@ const fullRefreshOption: INamedOption<yargs.Options> = {
     describe: "Forces incremental tables to be rebuilt from scratch.",
     type: "boolean",
     default: false
-  }
-};
-
-// It would be nice to use yargs' "implies" to implement this, but it doesn't work for some reason.
-const requiresSelection = (
-  name: string,
-  actions: INamedOption<yargs.Options>,
-  tags: INamedOption<yargs.Options>
-): INamedOption<yargs.Options>["check"] => (argv: yargs.Arguments) => {
-  if (argv[name] && !(argv[actions.name] || argv[tags.name])) {
-    throw new Error(
-      `The --${name} flag should only be supplied along with --${actions.name} or --${tags.name}.`
-    );
   }
 };
 
@@ -113,48 +97,6 @@ const includeDependentsOption: INamedOption<yargs.Options> = {
   check: requiresSelection("include-dependents", actionsOption, tagsOption)
 };
 
-// `compile` reuses the same prune() filtering as run/build, but these flags only
-// filter the *printed output* -- the whole project still compiles. The `output-`
-// prefix makes that distinction explicit.
-const outputActionsOption: INamedOption<yargs.Options> = {
-  name: "output-actions",
-  option: {
-    // No wildcard support: prune()'s matchPatterns() does exact matching on the
-    // action name or its fully-qualified `database.schema.name`.
-    describe: "A list of action names to filter the compiled output to.",
-    type: "array",
-    coerce: splitCommas
-  }
-};
-
-const outputTagsOption: INamedOption<yargs.Options> = {
-  name: "output-tags",
-  option: {
-    describe: "A list of tags to filter the compiled output to.",
-    type: "array",
-    coerce: splitCommas
-  }
-};
-
-const outputIncludeDepsOption: INamedOption<yargs.Options> = {
-  name: "output-include-deps",
-  option: {
-    describe: "If set, dependencies of the selected actions are also included in the output.",
-    type: "boolean"
-  },
-  check: requiresSelection("output-include-deps", outputActionsOption, outputTagsOption)
-};
-
-const outputIncludeDependentsOption: INamedOption<yargs.Options> = {
-  name: "output-include-dependents",
-  option: {
-    describe:
-      "If set, dependents (downstream) of the selected actions are also included in the output.",
-    type: "boolean"
-  },
-  check: requiresSelection("output-include-dependents", outputActionsOption, outputTagsOption)
-};
-
 const emitLineageOption: INamedOption<yargs.Options> = {
   name: "emit-lineage",
   option: {
@@ -163,21 +105,6 @@ const emitLineageOption: INamedOption<yargs.Options> = {
       "Overrides workflow_settings.yaml lineage.enabled when specified.",
     type: "boolean"
   }
-};
-
-const dotOutputOption: INamedOption<yargs.Options> = {
-  name: "dot",
-  option: {
-    describe: "Outputs a dot representation of the compiled project.",
-    type: "boolean",
-    default: false,
-  },
-    check: (argv: yargs.Arguments<any>) => {
-      if (argv.json && argv.dot) {
-        throw new Error("Arguments --json and --dot are mutually exclusive.");
-      }
-    }
-  
 };
 
 const executionTimeoutOption: INamedOption<yargs.Options> = {
@@ -238,9 +165,6 @@ const bigqueryJobLabelsOption: INamedOption<yargs.Options> = {
   }
 };
 
-const watchOptionName = "watch";
-
-const verboseOptionName = "verbose";
 const dryRunOptionName = "dry-run";
 const runTestsOptionName = "run-tests";
 
@@ -253,152 +177,7 @@ export function runCli() {
       initCommand,
       installCommand,
       initCredsCommand,
-      {
-        format: `compile [${projectDirMustExistOption.name}]`,
-        description:
-          "Compile the dataform project. Produces JSON output describing the non-executable graph.",
-        positionalOptions: [projectDirMustExistOption],
-        options: [
-          {
-            name: watchOptionName,
-            option: {
-              describe: "Whether to watch the changes in the project directory.",
-              type: "boolean",
-              default: false
-            }
-          },
-          jsonOutputOption,
-          dotOutputOption,
-          timeoutOption,
-          quietCompileOption,
-          outputActionsOption,
-          outputTagsOption,
-          outputIncludeDepsOption,
-          outputIncludeDependentsOption,
-          {
-            name: verboseOptionName,
-            option: {
-              describe: "Enable verbose compilation output. Example usage: 'dataform compile --verbose'",
-              type: "boolean",
-              default: false
-            },
-            check: (argv: yargs.Arguments) => {
-              if (argv.quiet && argv.verbose) {
-                throw new Error("Arguments --verbose and --quiet are mutually exclusive.");
-              }
-            }
-          },
-          ...ProjectConfigOptions.allYargsOptions
-        ],
-        processFn: async argv => {
-          const projectDir = argv[projectDirMustExistOption.name];
-          const logger = new Logger(!argv[jsonOutputOption.name]);
-
-          async function compileAndPrint() {
-
-            let outputType = compiledGraphOutputType.Summary;
-            if (argv[jsonOutputOption.name]) {
-              outputType = compiledGraphOutputType.Json;
-            } else if (argv[dotOutputOption.name]) {
-              outputType = compiledGraphOutputType.Dot;
-            }
-
-            if (outputType === compiledGraphOutputType.Summary) {
-              logger.log("Compiling...\n");
-            }
-            const compiledGraph = await compile({
-              projectDir,
-              projectConfigOverride: ProjectConfigOptions.constructProjectConfigOverride(argv),
-              timeoutMillis: argv[timeoutOption.name] || undefined,
-              verbose: argv[verboseOptionName] || false
-            });
-
-            // The whole project must compile (ref() resolution needs every action
-            // registered), but the printed output can be filtered to the selected
-            // action(s) -- mirroring how `run`/`build` prune the graph. We only prune
-            // a clean graph; if compilation produced errors we print the full graph
-            // plus the errors, keeping graph-level errors as-is.
-            const hasSelector =
-              argv[outputActionsOption.name]?.length > 0 || argv[outputTagsOption.name]?.length > 0;
-            const outputGraph =
-              hasSelector && !compiledGraphHasErrors(compiledGraph)
-                ? prune(compiledGraph, {
-                    actions: argv[outputActionsOption.name],
-                    tags: argv[outputTagsOption.name],
-                    includeDependencies: argv[outputIncludeDepsOption.name],
-                    includeDependents: argv[outputIncludeDependentsOption.name]
-                  })
-                : compiledGraph;
-            printCompiledGraph(outputGraph, outputType, argv[quietCompileOption.name]);
-            if (compiledGraphHasErrors(compiledGraph)) {
-              print("");
-              printCompiledGraphErrors(compiledGraph.graphErrors, argv[quietCompileOption.name]);
-              return true;
-            }
-            return false;
-          }
-
-          const graphHasErrors = await compileAndPrint();
-
-          if (!argv[watchOptionName]) {
-            return graphHasErrors ? 1 : 0;
-          }
-
-          let watching = true;
-
-          let timeoutID: NodeJS.Timer = null;
-          let isCompiling = false;
-
-          // Initialize watcher.
-          const watcher = chokidar.watch(projectDir, {
-            ignored: /node_modules/,
-            persistent: true,
-            ignoreInitial: true,
-            awaitWriteFinish: {
-              stabilityThreshold: 1000,
-              pollInterval: 200
-            }
-          });
-
-          const printReady = () => {
-            print("\nWatching for changes...\n");
-          };
-          // Add event listeners.
-          watcher
-            .on("ready", printReady)
-            .on("error", error => {
-              // This error is caught not if there is a compilation error, but
-              // if the watcher fails; this indicates an failure on our side.
-              printError(`Error: ${error}`);
-              process.exit(1);
-            })
-            .on("all", () => {
-              if (timeoutID || isCompiling) {
-                // don't recompile many times if we changed a lot of files
-                clearTimeout(timeoutID);
-              }
-
-              timeoutID = setTimeout(async () => {
-                clearTimeout(timeoutID);
-
-                if (!isCompiling) {
-                  isCompiling = true;
-                  await compileAndPrint();
-                  printReady();
-                  isCompiling = false;
-                }
-              }, RECOMPILE_DELAY);
-            });
-          process.on("SIGINT", async () => {
-            await watcher.close();
-            watching = false;
-            process.exit(1);
-          });
-          while (watching) {
-            await new Promise((resolve, reject) => setTimeout(() => resolve(), 100));
-          }
-        }
-      },
+      compileCommand,
       testCommand,
       {
         format: `run [${projectDirMustExistOption.name}]`,
